Move progress endpoints to progress.go
